# ports/openapi_server.gen.go
// Package ports provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.11.0 DO NOT EDIT.
package ports

import (
	"fmt"
	"net/http"

	"github.com/deepmap/oapi-codegen/pkg/runtime"
	"github.com/go-chi/chi/v5"
)

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Retrieve all channels
	// (GET /v1/channels)
	GetChannels(w http.ResponseWriter, r *http.Request)
	// Create a channel
	// (POST /v1/channels)
	CreateChannel(w http.ResponseWriter, r *http.Request)
	// Delete a channel
	// (DELETE /v1/channels/{channelID})
	DeleteChannelById(w http.ResponseWriter, r *http.Request, channelID string)
	// Retrieve a channel
	// (GET /v1/channels/{channelID})
	GetChannelById(w http.ResponseWriter, r *http.Request, channelID string)
	// Create a membership
	// (POST /v1/channels/{channelID}/memberships)
	CreateMembership(w http.ResponseWriter, r *http.Request, channelID string)
	// Remove a membership
	// (DELETE /v1/channels/{channelID}/memberships/{membershipID})
	DeleteMembershipById(w http.ResponseWriter, r *http.Request, channelID string, membershipID string)
	// Retrieve a membership
	// (GET /v1/channels/{channelID}/memberships/{membershipID})
	GetMembershipById(w http.ResponseWriter, r *http.Request, channelID string, membershipID string)
	// Opens a WebSocket connection
	// (GET /v1/stream)
	Websocket(w http.ResponseWriter, r *http.Request)
	// Retrieve all users
	// (GET /v1/users)
	GetUsers(w http.ResponseWriter, r *http.Request)
	// Create a user
	// (POST /v1/users)
	CreateUser(w http.ResponseWriter, r *http.Request)
	// Delete a user
	// (DELETE /v1/users/{userID})
	DeleteUserById(w http.ResponseWriter, r *http.Request, userID string)
	// Retrieve a user
	// (GET /v1/users/{userID})
	GetUserById(w http.ResponseWriter, r *http.Request, userID string)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.HandlerFunc) http.HandlerFunc

// GetChannels operation middleware
func (siw *ServerInterfaceWrapper) GetChannels(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	var handler = func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetChannels(w, r)
	}

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler(w, r.WithContext(ctx))
}

// CreateChannel operation middleware
func (siw *ServerInterfaceWrapper) CreateChannel(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	var handler = func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CreateChannel(w, r)
	}

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler(w, r.WithContext(ctx))
}

// DeleteChannelById operation middleware
func (siw *ServerInterfaceWrapper) DeleteChannelById(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	var err error

	// ------------- Path parameter "channelID" -------------
	var channelID string

	err = runtime.BindStyledParameter("simple", false, "channelID", chi.URLParam(r, "channelID"), &channelID)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "channelID", Err: err})
		return
	}

	var handler = func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.DeleteChannelById(w, r, channelID)
	}

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler(w, r.WithContext(ctx))
}

// GetChannelById operation middleware
func (siw *ServerInterfaceWrapper) GetChannelById(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	var err error

	// ------------- Path parameter "channelID" -------------
	var channelID string

	err = runtime.BindStyledParameter("simple", false, "channelID", chi.URLParam(r, "channelID"), &channelID)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "channelID", Err: err})
		return
	}

	var handler = func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetChannelById(w, r, channelID)
	}

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler(w, r.WithContext(ctx))
}

// CreateMembership operation middleware
func (siw *ServerInterfaceWrapper) CreateMembership(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	var err error

	// ------------- Path parameter "channelID" -------------
	var channelID string

	err = runtime.BindStyledParameter("simple", false, "channelID", chi.URLParam(r, "channelID"), &channelID)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "channelID", Err: err})
		return
	}

	var handler = func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CreateMembership(w, r, channelID)
	}

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler(w, r.WithContext(ctx))
}

// DeleteMembershipById operation middleware
func (siw *ServerInterfaceWrapper) DeleteMembershipById(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	var err error

	// ------------- Path parameter "channelID" -------------
	var channelID string

	err = runtime.BindStyledParameter("simple", false, "channelID", chi.URLParam(r, "channelID"), &channelID)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "channelID", Err: err})
		return
	}

	// ------------- Path parameter "membershipID" -------------
	var membershipID string

	err = runtime.BindStyledParameter("simple", false, "membershipID", chi.URLParam(r, "membershipID"), &membershipID)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "membershipID", Err: err})
		return
	}

	var handler = func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.DeleteMembershipById(w, r, channelID, membershipID)
	}

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler(w, r.WithContext(ctx))
}

// GetMembershipById operation middleware
func (siw *ServerInterfaceWrapper) GetMembershipById(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	var err error

	// ------------- Path parameter "channelID" -------------
	var channelID string

	err = runtime.BindStyledParameter("simple", false, "channelID", chi.URLParam(r, "channelID"), &channelID)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "channelID", Err: err})
		return
	}

	// ------------- Path parameter "membershipID" -------------
	var membershipID string

	err = runtime.BindStyledParameter("simple", false, "membershipID", chi.URLParam(r, "membershipID"), &membershipID)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "membershipID", Err: err})
		return
	}

	var handler = func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetMembershipById(w, r, channelID, membershipID)
	}

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler(w, r.WithContext(ctx))
}

// Websocket operation middleware
func (siw *ServerInterfaceWrapper) Websocket(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	var handler = func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.Websocket(w, r)
	}

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler(w, r.WithContext(ctx))
}

// GetUsers operation middleware
func (siw *ServerInterfaceWrapper) GetUsers(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	var handler = func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetUsers(w, r)
	}

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler(w, r.WithContext(ctx))
}

// CreateUser operation middleware
func (siw *ServerInterfaceWrapper) CreateUser(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	var handler = func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CreateUser(w, r)
	}

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler(w, r.WithContext(ctx))
}

// DeleteUserById operation middleware
func (siw *ServerInterfaceWrapper) DeleteUserById(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	var err error

	// ------------- Path parameter "userID" -------------
	var userID string

	err = runtime.BindStyledParameter("simple", false, "userID", chi.URLParam(r, "userID"), &userID)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "userID", Err: err})
		return
	}

	var handler = func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.DeleteUserById(w, r, userID)
	}

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler(w, r.WithContext(ctx))
}

// GetUserById operation middleware
func (siw *ServerInterfaceWrapper) GetUserById(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	var err error

	// ------------- Path parameter "userID" -------------
	var userID string

	err = runtime.BindStyledParameter("simple", false, "userID", chi.URLParam(r, "userID"), &userID)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "userID", Err: err})
		return
	}

	var handler = func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetUserById(w, r, userID)
	}

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler(w, r.WithContext(ctx))
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/v1/channels", wrapper.GetChannels)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/v1/channels", wrapper.CreateChannel)
	})
	r.Group(func(r chi.Router) {
		r.Delete(options.BaseURL+"/v1/channels/{channelID}", wrapper.DeleteChannelById)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/v1/channels/{channelID}", wrapper.GetChannelById)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/v1/channels/{channelID}/memberships", wrapper.CreateMembership)
	})
	r.Group(func(r chi.Router) {
		r.Delete(options.BaseURL+"/v1/channels/{channelID}/memberships/{membershipID}", wrapper.DeleteMembershipById)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/v1/channels/{channelID}/memberships/{membershipID}", wrapper.GetMembershipById)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/v1/stream", wrapper.Websocket)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/v1/users", wrapper.GetUsers)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/v1/users", wrapper.CreateUser)
	})
	r.Group(func(r chi.Router) {
		r.Delete(options.BaseURL+"/v1/users/{userID}", wrapper.DeleteUserById)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/v1/users/{userID}", wrapper.GetUserById)
	})

	return r
}